+++ lib/ecto/query/normalizer.ex
@@ -71,27 +71,27 @@ defmodule Ecto.Query.Normalizer do
 
   # Group by all fields
   defp normalize_group_by(query) do
-    entities = normalize_models(query.group_bys, query.sources)
-    %{query | group_bys: entities}
+    exprs = normalize_exprs(query.group_bys, query.sources)
+    %{query | group_bys: exprs}
   end
 
   # Add distinct on all field when model is in field list
   defp normalize_distinct(query) do
-    entities = normalize_models(query.distincts, query.sources)
-    %{query | distincts: entities}
+    exprs = normalize_exprs(query.distincts, query.sources)
+    %{query | distincts: exprs}
   end
 
-  # Expand model into all of its fields in an expression
-  defp normalize_models(query_expr, sources) do
+  # Expand var into all of its fields in an expression
+  defp normalize_exprs(query_expr, sources) do
     Enum.map(query_expr, fn expr ->
       new_expr =
         Enum.flat_map(expr.expr, fn
           {:&, _, _} = var ->
             model = Util.find_source(sources, var) |> Util.model
             fields = model.__schema__(:field_names)
-            Enum.map(fields, &{var, &1})
-          field ->
-            [field]
+            Enum.map(fields, &{{:., [], [var, &1]}, [], []})
+          expr ->
+            [expr]
         end)
       %{expr | expr: new_expr}
     end)