+++ test/ecto/query/validator_test.exs
@@ -208,6 +208,9 @@ defmodule Ecto.Query.ValidatorTest do
     query = Post |> distinct([p], p.id)
     validate(query)
 
+    query = Post |> distinct([p], p.id * 2)
+    validate(query)
+
     query = Post |> distinct([p], [p.id, p.title])
     validate(query)
 
@@ -261,7 +264,7 @@ defmodule Ecto.Query.ValidatorTest do
     validate(query)
 
     query = Post |> having([p], p.id) |> select([], 0)
-    assert_raise Ecto.QueryError, ~r"`Ecto.Query.ValidatorTest.Post.id` must appear in `group_by`", fn ->
+    assert_raise Ecto.QueryError, ~r"`&0.id\(\)` must appear in `group_by`", fn ->
       validate(query)
     end
   end
@@ -271,7 +274,7 @@ defmodule Ecto.Query.ValidatorTest do
     validate(query)
 
     query = Post |> group_by([p], p.id) |> having([p], p.title) |> select([], 0)
-    assert_raise Ecto.QueryError, ~r"`Ecto.Query.ValidatorTest.Post.title` must appear in `group_by`", fn ->
+    assert_raise Ecto.QueryError, ~r"`&0.title\(\)` must appear in `group_by`", fn ->
       validate(query)
     end
   end
@@ -280,8 +283,16 @@ defmodule Ecto.Query.ValidatorTest do
     query = Post |> group_by([p], p.id) |> select([p], p.id)
     validate(query)
 
+    query = Post |> group_by([p], p.id * 2) |> select([p], p.id * 2)
+    validate(query)
+
+    query = Post |> group_by([p], p.id * 2) |> select([p], p.id)
+    assert_raise Ecto.QueryError, ~r"`&0.id\(\)` must appear in `group_by`", fn ->
+      validate(query)
+    end
+
     query = Post |> group_by([p], p.id) |> select([p], p.title)
-    assert_raise Ecto.QueryError, ~r"`Ecto.Query.ValidatorTest.Post.title` must appear in `group_by`", fn ->
+    assert_raise Ecto.QueryError, ~r"`&0.title\(\)` must appear in `group_by`", fn ->
       validate(query)
     end
   end
@@ -291,7 +302,7 @@ defmodule Ecto.Query.ValidatorTest do
     validate(query)
 
     query = Post |> group_by([p], p.id) |> select([p], p)
-    assert_raise Ecto.QueryError, ~r"`Ecto.Query.ValidatorTest.Post.title` must appear in `group_by`", fn ->
+    assert_raise Ecto.QueryError, ~r"`&0.title\(\)` must appear in `group_by`", fn ->
       validate(query)
     end
   end